Fixed inventory item updates not working because we still had legacy product and inventoryitem confused
