+++ src/components/ProductCard.tsx
@@ -1,6 +1,6 @@
+import React from 'react';
 import { View, Pressable } from 'react-native';
 import {
-  TextInput as PaperTextInput,
   Card,
   IconButton,
   Text,
@@ -9,7 +9,7 @@ import {
 import { useNavigation } from '@react-navigation/native';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 
-import { Product } from '../database/database';
+import { Product } from '../database/models';
 import { RootStackParamList } from '../types/navigation';
 import { createHomeScreenStyles } from '../styles/HomeScreenStyles';
 import { getEmojiForProduct } from '../utils/stringUtils';
@@ -35,14 +35,7 @@ export const ProductCard = ({
   const theme = useTheme();
   const styles = createHomeScreenStyles(theme);
 
-  const {
-    quantity,
-    updateInventoryItemQuantity,
-    confirmRemoveProduct,
-    startContinuousAdjustment,
-    stopContinuousAdjustment,
-  } = useProduct({ productId: item.id, initialQuantity: item.quantity });
-
+  const { confirmRemoveProduct } = useProduct({ productId: item.id });
 
   return (
     <Card style={[styles.card, { opacity: isActive ? 0.5 : 1 }]}>
@@ -74,41 +67,6 @@ export const ProductCard = ({
                 iconColor={theme.colors.error}
                 testID={`delete-button-${item.id}`}
               />
-            </View>
-          </View>
-          <View style={styles.cardContent}>
-            <View style={styles.quantityContainer}>
-              <View style={styles.quantityInputContainer}>
-                <Text variant="bodyMedium">Quantidade: </Text>
-                <PaperTextInput
-                  mode="outlined"
-                  dense
-                  value={quantity.toString()}
-                  onChangeText={(value) =>
-                    updateInventoryItemQuantity(value === '' ? 0 : parseInt(value, 10))
-                  }
-                  keyboardType="numeric"
-                  style={styles.input}
-                  testID={`quantity-text-input-${item.id}`}
-                />
-              </View>
-              <View style={styles.quantityButtons}>
-                <IconButton
-                  icon="minus"
-                  size={20}
-                  onPress={() => updateInventoryItemQuantity(Math.max(0, quantity - 1))}
-                  onLongPress={() => startContinuousAdjustment(false)}
-                  onPressOut={stopContinuousAdjustment}
-                />
-                <IconButton
-                  icon="plus"
-                  size={20}
-                  onPress={() => updateInventoryItemQuantity(quantity + 1)}
-                  onLongPress={() => startContinuousAdjustment(true)}
-                  onPressOut={stopContinuousAdjustment}
-                  testID={`increment-button-${item.id}`}
-                />
-              </View>
             </View>
           </View>
         </Card.Content>
